feat(image-viewer-web): make sure the lightbox image does not include the thumbnail URL param
